Added comparison figure.
Current have 2 versions of render_part, one of which is better at
edges in the comparison figure.

Also upped the tolerance of the OpenCascade version of tesselate
to avoid problems that appear locally, but not on the servers.
The only cause I can see for this is that the servers use version
7.6.0, whereas I have 7.6.2 locally.