Model loading and complex part optimization attempt
Documented model loading formats in figures.ipynb

Tried to optimize a more complex part. Runs into several
issues; primarily, the radius of a cylinder goes negative,
resulting in non-realizable geometry. Additionally, when
non-plane geometry is invoved, it is possible for entities
not to intersect, so reconstruction fails (uncertain if this
is the actual problem we run into, however.)